setted probability to double to second decimal cifre

+++ HSA/src/com/hsa/adapter/PartialTextualAggregationAdapter.java
@@ -64,7 +64,7 @@ public class PartialTextualAggregationAdapter extends BaseAdapter{
 
         PartialTextualAggregation partialTextualAggregation = (PartialTextualAggregation)getItem(i);
         trackCost.setText("Cost: " + Integer.toString(partialTextualAggregation.getCost()));
-        trackProb.setText(Integer.toString(partialTextualAggregation.getProbability()) + "%" + "\n");
+        trackProb.setText(Double.toString(partialTextualAggregation.getProbability()) + "%" + "\n");
         trackProb.setTypeface(null, Typeface.BOLD);
         String name = "";
         String[] splitName = partialTextualAggregation.getName().split(" ");
